repositories

+++ backend/database/models/models.go
@@ -22,20 +22,20 @@ type User struct {
 
 // Post model
 type Post struct {
-	ID          int       `json:"id"`
-	UserID      int       `json:"user_id"`
+	ID          int64       `json:"id"`
+	UserID      int64       `json:"user_id"`
 	Content     string    `json:"content"`
 	ImagePath   *string   `json:"image_path,omitempty"`
-	PrivacyType int       `json:"privacy_type"` // 0: public, 1: friend, 2: private
+	PrivacyType int64       `json:"privacy_type"` // 0: public, 1: friend, 2: private
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
 
 // Comment model
 type Comment struct {
-	ID        int       `json:"id"`
-	PostID    int       `json:"post_id"`
-	UserID    int       `json:"user_id"`
+	ID        int64       `json:"id"`
+	PostID    int64       `json:"post_id"`
+	UserID    int64       `json:"user_id"`
 	Content   string    `json:"content"`
 	ImagePath *string   `json:"image_path"`
 	CreatedAt time.Time `json:"created_at"`
@@ -44,18 +44,18 @@ type Comment struct {
 
 // Follower model
 type Follower struct {
-	FollowerID  int       `json:"follower_id"`
-	FollowedID  int       `json:"followed_id"`
+	FollowerID  int64       `json:"follower_id"`
+	FollowedID  int64       `json:"followed_id"`
 	Accepted    bool      `json:"accepted"`
 	FollowedAt  time.Time `json:"followed_at"`
 }
 
 // Message model
 type Message struct {
-	ID         int        `json:"id"`
-	SenderID   int        `json:"sender_id"`
-	ReceiverID int        `json:"receiver_id"`
-	GroupID    *int       `json:"group_id"`
+	ID         int64        `json:"id"`
+	SenderID   int64        `json:"sender_id"`
+	ReceiverID int64        `json:"receiver_id"`
+	GroupID    *int64       `json:"group_id"`
 	Content    string     `json:"content"`
 	CreatedAt  time.Time  `json:"created_at"`
 	ReadAt     *time.Time `json:"read_at"`
@@ -63,20 +63,20 @@ type Message struct {
 
 //Notification model
 type Notification struct {
-	ID            int       `json:"id"`
-	UserID        int       `json:"user_id"`
+	ID            int64       `json:"id"`
+	UserID        int64       `json:"user_id"`
 	Type          string    `json:"type"`
 	Content       string    `json:"content"`
 	Read          bool      `json:"read"`
-	ReferenceID   *int      `json:"reference_id"`
+	ReferenceID   *int64      `json:"reference_id"`
 	ReferenceType *string   `json:"reference_type"`
 	CreatedAt     time.Time `json:"created_at"`
 }
 
 // Session model
 type Session struct {
-	ID           int       `json:"id"`
-	UserID       int       `json:"user_id"`
+	ID           int64       `json:"id"`
+	UserID       int64       `json:"user_id"`
 	SessionToken string    `json:"session_token"`
 	CreatedAt    time.Time `json:"created_at"`
 	ExpiresAt    time.Time `json:"expires_at"`
@@ -86,8 +86,8 @@ type Session struct {
 
 // Group model
 type Group struct {
-	ID          int       `json:"id"`
-	CreatorID   int       `json:"creator_id"`
+	ID          int64       `json:"id"`
+	CreatorID   int64       `json:"creator_id"`
 	Title       string    `json:"title"`
 	Description *string   `json:"description"`
 	CreatedAt   time.Time `json:"created_at"`
@@ -96,36 +96,36 @@ type Group struct {
 
 // GroupPrivacy model
 type PostPrivacy struct {
-	ID           int    `json:"id"`
-	PostID       int    `json:"post_id"`
-	UserID       int    `json:"user_id"`
+	ID           int64    `json:"id"`
+	PostID       int64    `json:"post_id"`
+	UserID       int64    `json:"user_id"`
 	PrivacyLevel string `json:"privacy_level"` // "public", "followers", "private"
 }
 
 // GroupMember model
 type GroupMember struct {
-	ID        int       `json:"id"`
-	GroupID   int       `json:"group_id"`
-	UserID    int       `json:"user_id"`
+	ID        int64       `json:"id"`
+	GroupID   int64       `json:"group_id"`
+	UserID    int64       `json:"user_id"`
 	Accepted  bool      `json:"accepted"`
 	CreatedAt time.Time `json:"create_at"`
 }
 
 // GroupInvitation model
 type GroupInvitation struct {
-	ID         int       `json:"id"`
-	GroupID    int       `json:"group_id"`
-	InviterID  int       `json:"inviter_id"`
-	InviteeID  int       `json:"invitee_id"`
+	ID         int64       `json:"id"`
+	GroupID    int64       `json:"group_id"`
+	InviterID  int64       `json:"inviter_id"`
+	InviteeID  int64       `json:"invitee_id"`
 	Pending    bool      `json:"pending"`
 	CreatedAt  time.Time `json:"create_at"`
 }
 
 // GroupPost model
 type GroupPost struct {
-	ID        int       `json:"id"`
-	GroupID   int       `json:"group_id"`
-	UserID    int       `json:"user_id"`
+	ID        int64       `json:"id"`
+	GroupID   int64       `json:"group_id"`
+	UserID    int64       `json:"user_id"`
 	Content   string    `json:"content"`
 	ImagePath *string   `json:"image_path"`
 	CreatedAt time.Time `json:"create_at"`
@@ -134,9 +134,9 @@ type GroupPost struct {
 
 // GroupComment model
 type GroupComment struct {
-	ID          int       `json:"id"`
-	GroupPostID int       `json:"group_post_id"`
-	UserID      int       `json:"user_id"`
+	ID          int64       `json:"id"`
+	GroupPostID int64       `json:"group_post_id"`
+	UserID      int64       `json:"user_id"`
 	Content     string    `json:"content"`
 	CreatedAt   time.Time `json:"create_at"`
 	UpdatedAt   time.Time `json:"update_at"`
@@ -146,9 +146,9 @@ type GroupComment struct {
 
 // Event model
 type Event struct {
-	ID          int       `json:"id"`
-	GroupID     int       `json:"group_id"`
-	CreatorID   int       `json:"creator_id"`
+	ID          int64       `json:"id"`
+	GroupID     int64       `json:"group_id"`
+	CreatorID   int64       `json:"creator_id"`
 	Title       string    `json:"title"`
 	Description *string   `json:"description"`
 	EventDate   time.Time `json:"event_date"`
@@ -158,17 +158,17 @@ type Event struct {
 
 // EventOption model
 type EventOption struct {
-	ID        int    `json:"id"`
-	EventID   int    `json:"event_id"`
+	ID        int64    `json:"id"`
+	EventID   int64    `json:"event_id"`
 	OptionText string `json:"option_text"`
 }
 
 // EventResponse model
 type EventResponse struct {
-	ID        int        `json:"id"`
-	EventID   int        `json:"event_id"`
-	UserID    int        `json:"user_id"`
-	OptionID  *int       `json:"option_id"`
+	ID        int64        `json:"id"`
+	EventID   int64        `json:"event_id"`
+	UserID    int64        `json:"user_id"`
+	OptionID  *int64       `json:"option_id"`
 	CreatedAt time.Time  `json:"create_at"`
 }
